requirements.txt
pytest
pytest-playwright
playwright
pytest-html
colorlog
python-dotenv